refactor useDefaultMapOverlay
